# Speed up startup and reload times