@dwijnand 2017-03-15
2 more things: (1) I restarted the failure on Travis CI, and (2) nice work!