@retronym 2017-03-17
I'd suggest to drop the last commit for now. Parsing is really quite cheap once you've got the compiler initialized, so I don't think future proofing this code so as not to be a point of contention for a hypothetical multi-threaded project load is worth.