@jvican 2017-03-17
> I'd also suggest starting with an initial commit that reuses the ToolBox based parser. If for some reason, we have to revert the change to use Global directly (hypothetically this might happen if the compiler API changes in a future compiler that SBT upgrades to and this code is struggling to fina a maintainer), we'll still have the main performance improvement.
  
The problem is that toolbox is not thread-safe. You cannot use it in the same way I'm using it now because reporter errors will clash with each other. This is the reason why I'm doing all the magic with the `reporterId`s. I actually think that we may want to parse sbt files in parallel (if sbt is not already doing this). I heard somewhere that scalac was parsing in parallel too.
    
> Parsing is really quite cheap once you've got the compiler initialized
Isn't there an overhead of using `ReflectGlobal` over just `Global`? The current approach in this commit, as I see it, avoids class loading some classes.
  
True, parsing is quite cheap. The reason I went for this solution though is because I gained 400ms. I'll repeat my measurements with the global toolbox, but we still have the issue that is not thread-safe and we cannot instrument it.
  
However, a bigger benefit with this approach is that we can reuse the parsed trees from this concrete parser and feed them into eval, instead of parsing the whole compilation unit again. This will reduce quite some code and ease maintenance. Potentially, we could even reuse the global of eval for parsing, avoiding the initialization of the extra global, which I believe it may be non-negligible. WDYT?