@retronym 2017-03-17
The lack of thread safety is moot if we live with synchronisation. I
believe will still be a big win by avoiding unneeded init cost

I'm in favour of the end result of using global directly; but I think it's
worth an extra commit to make that a separate step.