@jvican 2017-03-17
@retronym So what you propose is that I include a commit before my changes to parser that uses a global toolbox instance and is synchronized, no?
  
I'm happy to do that, I have that commit lying around my computer somewhere.