@jvican 2017-04-18
I'm closing this because I'm reopening against 1.0.