@jvican 2017-04-18
Hello, Jason's changes have been applied in https://github.com/sbt/sbt/pull/3115, which is rebased to latest master 1.0.